+++ packages/google_mobile_ads/android/src/main/java/io/flutter/plugins/googlemobileads/FlutterInterstitialAd.java
@@ -73,11 +73,22 @@ class FlutterInterstitialAd extends FlutterAd.FlutterOverlayAd {
   @Override
   public void show() {
     if (ad == null) {
-      Log.e(TAG, "The interstitial wasn't loaded yet.");
+      Log.e(TAG, "Error showing interstitial - the interstitial ad wasn't loaded yet.");
       return;
     }
     ad.setFullScreenContentCallback(new FlutterFullScreenContentCallback(manager, adId));
     ad.show(manager.activity);
+  }
+
+  @Override
+  public void setImmersiveMode(boolean immersiveModeEnabled) {
+    if (ad == null) {
+      Log.e(
+          TAG,
+          "Error setting immersive mode in interstitial ad - the interstitial ad wasn't loaded yet.");
+      return;
+    }
+    ad.setImmersiveMode(immersiveModeEnabled);
   }
 
   /** An InterstitialAdLoadCallback that just forwards events to a delegate. */