+++ packages/google_mobile_ads/ios/Classes/FLTAdUtil.m
@@ -12,5 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-/** Versioned request agent string. */
-#define FLT_REQUEST_AGENT_VERSIONED @"Flutter-GMA-0.13.4"
+#import "FLTAdUtil.h"
+
+@implementation FLTAdUtil
+
++ (BOOL)isNull:(id)object {
+  return object == nil || [[NSNull null] isEqual:object];
+}
+
++ (BOOL)isNotNull:(id)object {
+  return ![FLTAdUtil isNull:object];
+}
+
+@end