+++ packages/google_mobile_ads/test/mobile_ads_test.dart
@@ -46,9 +46,14 @@ void main() {
               ),
             });
           case '_init':
-            return null;
           case 'MobileAds#setSameAppKeyEnabled':
+          case 'MobileAds#setAppMuted':
+          case 'MobileAds#setAppVolume':
+          case 'MobileAds#disableSDKCrashReporting':
+          case 'MobileAds#disableMediationInitialization':
             return null;
+          case 'MobileAds#getVersionString':
+            return Future<String>.value('Test-SDK-Version');
           default:
             assert(false);
             return null;
@@ -147,6 +152,131 @@ void main() {
             arguments: {'isEnabled': true}),
         isMethodCall('MobileAds#setSameAppKeyEnabled',
             arguments: {'isEnabled': false})
+      ]);
+    });
+
+    test('encode/decode empty native ad options', () {
+      debugDefaultTargetPlatformOverride = TargetPlatform.android;
+      ByteData byteData = codec.encodeMessage(NativeAdOptions())!;
+
+      NativeAdOptions result = codec.decodeMessage(byteData);
+      expect(result.mediaAspectRatio, null);
+      expect(result.adChoicesPlacement, null);
+      expect(result.requestCustomMuteThisAd, null);
+      expect(result.shouldRequestMultipleImages, null);
+      expect(result.shouldReturnUrlsForImageAssets, null);
+      expect(result.videoOptions, null);
+
+      byteData =
+          codec.encodeMessage(NativeAdOptions(videoOptions: VideoOptions()))!;
+      result = codec.decodeMessage(byteData);
+      expect(result.mediaAspectRatio, null);
+      expect(result.adChoicesPlacement, null);
+      expect(result.requestCustomMuteThisAd, null);
+      expect(result.shouldRequestMultipleImages, null);
+      expect(result.shouldReturnUrlsForImageAssets, null);
+      expect(result.videoOptions!.clickToExpandRequested, null);
+      expect(result.videoOptions!.customControlsRequested, null);
+      expect(result.videoOptions!.startMuted, null);
+    });
+
+    test('encode/decode native ad options', () {
+      NativeAdOptions nativeAdOptions = NativeAdOptions(
+          adChoicesPlacement: AdChoicesPlacement.topRightCorner,
+          mediaAspectRatio: MediaAspectRatio.unknown,
+          videoOptions: VideoOptions(
+            clickToExpandRequested: false,
+            customControlsRequested: false,
+            startMuted: false,
+          ),
+          requestCustomMuteThisAd: false,
+          shouldRequestMultipleImages: false,
+          shouldReturnUrlsForImageAssets: false);
+
+      debugDefaultTargetPlatformOverride = TargetPlatform.android;
+      ByteData byteData = codec.encodeMessage(nativeAdOptions)!;
+
+      NativeAdOptions result = codec.decodeMessage(byteData);
+      expect(result.mediaAspectRatio, MediaAspectRatio.unknown);
+      expect(result.adChoicesPlacement, AdChoicesPlacement.topRightCorner);
+      expect(result.requestCustomMuteThisAd, false);
+      expect(result.shouldRequestMultipleImages, false);
+      expect(result.shouldReturnUrlsForImageAssets, false);
+      expect(result.videoOptions!.startMuted, false);
+      expect(result.videoOptions!.customControlsRequested, false);
+      expect(result.videoOptions!.clickToExpandRequested, false);
+
+      nativeAdOptions = NativeAdOptions(
+          adChoicesPlacement: AdChoicesPlacement.bottomLeftCorner,
+          mediaAspectRatio: MediaAspectRatio.landscape,
+          videoOptions: VideoOptions(
+            clickToExpandRequested: true,
+            customControlsRequested: true,
+            startMuted: true,
+          ),
+          requestCustomMuteThisAd: true,
+          shouldRequestMultipleImages: true,
+          shouldReturnUrlsForImageAssets: true);
+
+      byteData = codec.encodeMessage(nativeAdOptions)!;
+      result = codec.decodeMessage(byteData);
+
+      expect(result.mediaAspectRatio, MediaAspectRatio.landscape);
+      expect(result.adChoicesPlacement, AdChoicesPlacement.bottomLeftCorner);
+      expect(result.requestCustomMuteThisAd, true);
+      expect(result.shouldRequestMultipleImages, true);
+      expect(result.shouldReturnUrlsForImageAssets, true);
+      expect(result.videoOptions!.startMuted, true);
+      expect(result.videoOptions!.customControlsRequested, true);
+      expect(result.videoOptions!.clickToExpandRequested, true);
+    });
+
+    test('$MobileAds.setAppMuted', () async {
+      await MobileAds.instance.setAppMuted(true);
+
+      expect(log, <Matcher>[
+        isMethodCall('MobileAds#setAppMuted', arguments: {'muted': true})
+      ]);
+
+      await MobileAds.instance.setAppMuted(false);
+
+      expect(log, <Matcher>[
+        isMethodCall('MobileAds#setAppMuted', arguments: {'muted': true}),
+        isMethodCall('MobileAds#setAppMuted', arguments: {'muted': false})
+      ]);
+    });
+
+    test('$MobileAds.setAppVolume', () async {
+      await MobileAds.instance.setAppVolume(1.0);
+
+      expect(log, <Matcher>[
+        isMethodCall('MobileAds#setAppVolume', arguments: {'volume': 1.0})
+      ]);
+    });
+
+    test('$MobileAds.disableSDKCrashReporting', () async {
+      debugDefaultTargetPlatformOverride = TargetPlatform.iOS;
+      await MobileAds.instance.disableSDKCrashReporting();
+
+      expect(log, <Matcher>[
+        isMethodCall('MobileAds#disableSDKCrashReporting', arguments: null)
+      ]);
+    });
+
+    test('$MobileAds.disableMediationInitialization', () async {
+      await MobileAds.instance.disableMediationInitialization();
+
+      expect(log, <Matcher>[
+        isMethodCall('MobileAds#disableMediationInitialization',
+            arguments: null)
+      ]);
+    });
+
+    test('$MobileAds.getVersionString', () async {
+      await MobileAds.instance.getVersionString();
+
+      expect(log, <Matcher>[
+        isMethodCall('MobileAds#getVersionString', arguments: null)
       ]);
     });
   });